32 -> 64 bit data types

+++ estimator.c
@@ -10,8 +10,8 @@ void tinysync_est_state_t_initialize(tinysync_est_state_t * state){
 }
 
 void calculate_line(tinysync_constraint_t p_1, tinysync_constraint_t p_2, tinysync_line_t * line){
-    line->a = (float)(p_2.t_1 - p_1.t_1) / (float)(p_2.t_2 - p_1.t_2);
-    line->b = (float)(p_1.t_1) - line->a * (float)(p_1.t_2);
+    line->a = (double)(p_2.t_1 - p_1.t_1) / (double)(p_2.t_2 - p_1.t_2);
+    line->b = (double)(p_1.t_1) - line->a * (double)(p_1.t_2);
 }
 
 //estimator of linear clock drift and offset
@@ -50,9 +50,9 @@ tinysync_est_ret_t tinysync_est_etimate(tinysync_est_state_t * state, const tiny
     default: // Algorithm initialized, normal behaviour
 
         // Handle new lower constraint (b_3)
-        if        (b_3.t_1 < state->lineset.ab.b + state->lineset.ab.a * (float)(b_3.t_2)){
+        if        (b_3.t_1 < state->lineset.ab.b + state->lineset.ab.a * (double)(b_3.t_2)){
             // Lower constraint too low to be useful: do nothing
-        } else if (b_3.t_1 > state->lineset.ba.b + state->lineset.ba.a * (float)(b_3.t_2)){
+        } else if (b_3.t_1 > state->lineset.ba.b + state->lineset.ba.a * (double)(b_3.t_2)){
             // Lower constraint too high to maintain linear model: remove old constraints and add new.
             a_1 = state->constraints.a_2;
             b_1 = state->constraints.b_2;
@@ -63,16 +63,16 @@ tinysync_est_ret_t tinysync_est_etimate(tinysync_est_state_t * state, const tiny
         } else {
             // Lower constraint useful: update b_2
             b_2 = b_3;
-            if ( (!tinysync_constraint_t_compare(state->constraints.a_1,state->constraints.a_2)) && (b_3.t_1 > state->lineset.aa.b + state->lineset.aa.a * (float)(b_3.t_2)) ){
+            if ( (!tinysync_constraint_t_compare(state->constraints.a_1,state->constraints.a_2)) && (b_3.t_1 > state->lineset.aa.b + state->lineset.aa.a * (double)(b_3.t_2)) ){
                 // a_2 is better than a_1 for this lower constraint: update a_1
                 a_1 = state->constraints.a_2;
             }
         }
 
         // Handle new upper constraint (a_3)
-        if        (a_3.t_1 > state->lineset.ba.b + state->lineset.ba.a * (float)(a_3.t_2)){
+        if        (a_3.t_1 > state->lineset.ba.b + state->lineset.ba.a * (double)(a_3.t_2)){
             // Upper constraint too high to be useful: do nothing
-        } else if (a_3.t_1 < state->lineset.ab.b + state->lineset.ab.a * (float)(a_3.t_2)){
+        } else if (a_3.t_1 < state->lineset.ab.b + state->lineset.ab.a * (double)(a_3.t_2)){
             // Upper constraint too low to maintain linear model: remove old constraints and add new.
             a_1 = state->constraints.a_2;
             b_1 = state->constraints.b_2;
@@ -83,7 +83,7 @@ tinysync_est_ret_t tinysync_est_etimate(tinysync_est_state_t * state, const tiny
         } else {
             // Upper constraint useful: update a_2
             a_2 = a_3;
-            if ( (!tinysync_constraint_t_compare(state->constraints.b_1,state->constraints.b_2)) && (a_3.t_1 < state->lineset.bb.b + state->lineset.bb.a * (float)(a_3.t_2)) ){
+            if ( (!tinysync_constraint_t_compare(state->constraints.b_1,state->constraints.b_2)) && (a_3.t_1 < state->lineset.bb.b + state->lineset.bb.a * (double)(a_3.t_2)) ){
                 // b_2 is better than b_1 for this upper constraint: update b_1
                 b_1 = state->constraints.b_2;
             }
